Move LoggingFacility to its own file
